feat: Refactor chat friend navigation
