+++ src/routes/(app)/chat/+layout.ts
@@ -1,7 +1,8 @@
-import { goto } from '$app/navigation';
 import { userData } from '$lib/mock/users';
-import * as cons from 'gun';
 import type { LayoutLoad } from './$types';
+import "@src/lib/initGun"
+import { selectedId } from './stores';
+import { get } from 'svelte/store';
 
 type FriendsMap = {
   [id: string]: User
@@ -15,14 +16,23 @@ function getLastOpenedFriend() {
   return "laracroft#4512"
 }
 
+function getSelectedFriend() {
+  return get(selectedId) || getLastOpenedFriend()
+}
+
+const me: User = {
+  profilePicture: "https://images.pexels.com/photos/220453/pexels-photo-220453.jpeg?cs=srgb&dl=pexels-pixabay-220453.jpg",
+  name: "Andrew McEuler",
+  id: "andrewmc#2718",
+  pubKey: ((typeof window !== "undefined") && JSON.parse(localStorage.getItem("currentUser")!).pub) || "TsRMyecmURD8iqH9A_qDRYDOfQp7xQN68m8z3QstG48.zWVVasKlG9E2Kx2wf1UFyEiphnAnyw8EyQXNGaP1uBI",
+  lastMessageTs: 1672718281828
+}
+
 export const load = (({ url }) => {
-  let lastOpenedFriend = getLastOpenedFriend()
-  if (url.pathname == "/chat") {
-    goto(`/chat/${lastOpenedFriend.replace(/#/g, '%23')}`)
-  }
-  let selectedFriend = (url.pathname.split("/")[2] || lastOpenedFriend).replace(/%23/g, '#')
+  let selected = getSelectedFriend()
   return {
     friends: getFriends(),
-    selected: selectedFriend
+    selected: selected,
+    me: me,
   };
 }) satisfies LayoutLoad;
