+++ src/routes/(app)/chat/stores.ts
@@ -1,0 +1,4 @@
+import {writable} from "svelte/store"
+
+export const activeCall = writable<string | null>(null)
+export const selectedId = writable<string | null>(null)
